Move `write.py` to main namespace, and extract `_write_elem_with_overwrite` into `tables/v1.py`
