+++ fractal_tasks_core/ome_zarr/label_group.py
@@ -6,7 +6,7 @@ import zarr.hierarchy
 from pydantic.error_wrappers import ValidationError
 
 from ..ngff import NgffImageMeta
-from .write import OverwriteNotAllowedError
+from ..write import OverwriteNotAllowedError
 
 
 def prepare_label_group(
