+++ fractal_tasks_core/tasks/maximum_intensity_projection.py
@@ -23,7 +23,7 @@ from zarr.errors import ContainsArrayError
 
 from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.write import OverwriteNotAllowedError
+from fractal_tasks_core.write import OverwriteNotAllowedError
 
 logger = logging.getLogger(__name__)
