+++ tests/tasks/test_workflows_napari_workflows.py
@@ -27,10 +27,10 @@ from ._validation import validate_schema
 from .lib_empty_ROI_table import _add_empty_ROI_table
 from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsInput
 from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsOutput
-from fractal_tasks_core.ome_zarr.write import OverwriteNotAllowedError
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
 )
+from fractal_tasks_core.write import OverwriteNotAllowedError
 
 try:
     import napari_skimage_regionprops_mock
